requirements.txt
google-generativeai
gradio>=4.32
python-dotenv
fpdf